Share one ManualSessionComposer between /sessions/new and the create-session modal
The composer card (textarea, mention/slash pickers, attachments, repo/branch/
model/reasoning controls, mobile settings sheet, drag-and-drop) was duplicated
across the page and the dialog, so the modal had drifted: its + menu was
missing the Linear-issue link and its model picker showed agents the user had
no credentials for. Extract the composer into a single component used by both
surfaces, so future composer work lands in one place. Drafts are enabled in
both, with the dialog mounting the composer only while open so reopens reset
state cleanly.
