+++ frontend/src/components/create-session-dialog.test.tsx
@@ -115,15 +115,15 @@ describe("CreateSessionDialog", () => {
     ).toBeInTheDocument();
   });
 
-  it("disables Create button when textarea is empty", () => {
+  it("disables Start session button when textarea is empty", () => {
     renderWithProviders(
       <CreateSessionDialog open onOpenChange={onOpenChange} />,
     );
 
-    expect(screen.getByRole("button", { name: /Create/ })).toBeDisabled();
+    expect(screen.getByRole("button", { name: "Start session" })).toBeDisabled();
   });
 
-  it("enables Create button when textarea has text", async () => {
+  it("enables Start session button when textarea has text", async () => {
     const user = userEvent.setup();
 
     renderWithProviders(
@@ -135,7 +135,7 @@ describe("CreateSessionDialog", () => {
       "Fix the login bug",
     );
 
-    expect(screen.getByRole("button", { name: /Create/ })).toBeEnabled();
+    expect(screen.getByRole("button", { name: "Start session" })).toBeEnabled();
   });
 
   it("calls onOpenChange(false) on successful submission", async () => {
@@ -151,7 +151,7 @@ describe("CreateSessionDialog", () => {
       "Fix the login bug",
     );
 
-    await user.click(screen.getByRole("button", { name: /Create/ }));
+    await user.click(screen.getByRole("button", { name: "Start session" }));
 
     await waitFor(() => {
       expect(onOpenChange).toHaveBeenCalledWith(false);
@@ -179,7 +179,7 @@ describe("CreateSessionDialog", () => {
       "Fix the login bug",
     );
 
-    await user.click(screen.getByRole("button", { name: /Create/ }));
+    await user.click(screen.getByRole("button", { name: "Start session" }));
 
     await waitFor(() => {
       expect(screen.getByText(/Something went wrong/)).toBeInTheDocument();
@@ -531,7 +531,7 @@ describe("CreateSessionDialog", () => {
       });
 
       // Submit
-      await user.click(screen.getByRole("button", { name: /Create/ }));
+      await user.click(screen.getByRole("button", { name: "Start session" }));
 
       await waitFor(() => {
         expect(onOpenChange).toHaveBeenCalledWith(false);
@@ -563,6 +563,14 @@ describe("CreateSessionDialog", () => {
           },
         });
       }),
+      // Gemini must be available in the model picker for this test to be able
+      // to select "gemini-2.5-pro"; the composer hides agents the user has no
+      // credentials for.
+      http.get("/api/v1/settings/credentials/resolved", () => {
+        return HttpResponse.json({
+          data: [{ provider: "gemini", source: "personal" }],
+        });
+      }),
       http.post("/api/v1/sessions/manual", async ({ request }) => {
         requestBody = await request.json() as Record<string, unknown>;
         return HttpResponse.json({
@@ -586,7 +594,7 @@ describe("CreateSessionDialog", () => {
       screen.getByPlaceholderText("Tell the agent what to do..."),
       "Fix the login bug",
     );
-    await user.click(screen.getByRole("button", { name: /Create/ }));
+    await user.click(screen.getByRole("button", { name: "Start session" }));
 
     await waitFor(() => {
       expect(onOpenChange).toHaveBeenCalledWith(false);
